+++ LessonMonitor/LessonMonitor.BL/GitHubService.cs
@@ -1,0 +1,8 @@
+﻿using LessonMonitor.Core;
+
+namespace LessonMonitor.BL
+{
+    internal class GitHubService: IGitHubService
+    {
+    }
+}
