+++ LessonMonitor/LessonMonitor.BL/LessonMonitor.BL.csproj
@@ -1,0 +1,14 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\LessonMonitor.Core\LessonMonitor.Core.csproj" />
+    <ProjectReference Include="..\LessonMonitor.Domain\LessonMonitor.Domain.csproj" />
+  </ItemGroup>
+
+</Project>
