+++ LessonMonitor/LessonMonitor.BL/UserRespository.cs
@@ -1,0 +1,13 @@
+﻿using LessonMonitor.Core;
+using LessonMonitor.Domain;
+
+namespace LessonMonitor.BL
+{
+    public class UserRespository : IUserRepository
+    {
+        public User Get(string userName)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
