+++ LessonMonitor/LessonMonitor.Core/ICacheManager.cs
@@ -1,0 +1,6 @@
+﻿namespace LessonMonitor.Core
+{
+    public interface ICacheManager
+    {
+    }
+}
