+++ LessonMonitor/LessonMonitor.Core/IGitHubService.cs
@@ -1,0 +1,6 @@
+﻿namespace LessonMonitor.Core
+{
+    public interface IGitHubService
+    {
+    }
+}
