+++ LessonMonitor/LessonMonitor.sln
@@ -1,13 +1,19 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30804.86
+# Visual Studio Version 17
+VisualStudioVersion = 17.1.32414.318
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LessonMonitor.API", "LessonMonitor.API\LessonMonitor.API.csproj", "{2D5C21A9-2B86-4E4B-B751-9A9F16D7DFB5}"
 EndProject
 Project("{00D1A9C2-B5F0-4AF3-8072-F6C62B433612}") = "LessonMonitor.DataBase", "LessonMonitor.DataBase\LessonMonitor.DataBase.sqlproj", "{A54EC47F-1ECB-4188-A127-CB465FB67979}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConsoleApp1", "ConsoleApp1\ConsoleApp1.csproj", "{DE2AA10A-F3C0-444A-8211-42709A44F396}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConsoleApp1", "ConsoleApp1\ConsoleApp1.csproj", "{DE2AA10A-F3C0-444A-8211-42709A44F396}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LessonMonitor.Core", "LessonMonitor.Core\LessonMonitor.Core.csproj", "{4867DC94-4AD9-4F63-8181-8DFE1E81873A}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LessonMonitor.Domain", "LessonMonitor.Domain\LessonMonitor.Domain.csproj", "{9753EBBB-2BC0-44AD-B380-FD14B69420A3}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LessonMonitor.BL", "LessonMonitor.BL\LessonMonitor.BL.csproj", "{6F99F690-CC80-4D99-AF94-F023127A56EB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -29,6 +35,18 @@ Global
 		{DE2AA10A-F3C0-444A-8211-42709A44F396}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DE2AA10A-F3C0-444A-8211-42709A44F396}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DE2AA10A-F3C0-444A-8211-42709A44F396}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4867DC94-4AD9-4F63-8181-8DFE1E81873A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4867DC94-4AD9-4F63-8181-8DFE1E81873A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4867DC94-4AD9-4F63-8181-8DFE1E81873A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4867DC94-4AD9-4F63-8181-8DFE1E81873A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9753EBBB-2BC0-44AD-B380-FD14B69420A3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9753EBBB-2BC0-44AD-B380-FD14B69420A3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9753EBBB-2BC0-44AD-B380-FD14B69420A3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9753EBBB-2BC0-44AD-B380-FD14B69420A3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6F99F690-CC80-4D99-AF94-F023127A56EB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6F99F690-CC80-4D99-AF94-F023127A56EB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6F99F690-CC80-4D99-AF94-F023127A56EB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6F99F690-CC80-4D99-AF94-F023127A56EB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
